pyproject.toml
[project]
name = "footScrapper"
version = "1.0.0"
description = "Python command line interface to list upcoming matchs for given team (using allsportsapi.com)"
dependencies = [
  "docopt",
  "human-readable>=1.3.4",
  "python-dotenv",
  "pytz>=2024.2",
  "requests>=2.32.3",
  "tzlocal>=5.2",
]
requires-python = ">=3.8"

[tool.ruff]
line-length = 110
